Remove example SSH key
People using the quickstart might not be aware that this key is public
and needs to be regenerated for a proper production setup. It's easier
to remove the key directly to avoid future misuse. Instead, a script is
provided to generate keys on demand.

Also added ecdsa and ed25519 keys additionally.